backend/requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.1
pydantic-settings>=2.2.1

# Google Gemini (LLM + embeddings)
google-generativeai>=0.8.3

# Database
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.10
asyncpg>=0.29.0
pgvector>=0.2.5

# Document parsing
pypdf>=4.2.0
PyMuPDF>=1.23.0
python-pptx>=0.6.23
python-docx>=1.1.2
pandas>=2.2.2
openpyxl>=3.1.0
Pillow>=10.0.0

# Orchestration (LangGraph state machine)
langgraph>=0.2.0
langchain-core>=0.3.0

# Utilities
numpy>=1.26.4
python-dotenv==1.0.1
